websockets/requirements.txt
rel==0.4.9
websocket-client==1.6.4
